requirements.txt
requests==2.27.1
feedgen==0.9.0
PyGithub==1.55
cryptography==36.0.2
PyJWT==2.3.0
python-telegram-bot==13.11
